Move DocumentUploadViewer and DocumentContent to /shared/DocumentViewer/ directory
Conflicts:
	src/scenes/Office/DocumentViewer/index.jsx
