+++ src/scenes/Office/OrdersInfo.jsx
@@ -6,8 +6,8 @@ import { compact, get } from 'lodash';
 
 import LoadingPlaceholder from 'shared/LoadingPlaceholder';
 import Alert from 'shared/Alert';
+import DocumentContent from 'shared/DocumentViewer/DocumentContent';
 import OrdersViewerPanel from './OrdersViewerPanel';
-import DocumentContent from './DocumentViewer/DocumentContent';
 import { loadMoveDependencies } from './ducks.js';
 
 import './office.css';
